+++ packages/util-crypto/src/address/index.ts
@@ -3,12 +3,14 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import checkAddress from './check';
+import checkAddressChecksum from './checkChecksum';
 import decodeAddress from './decode';
 import encodeAddress from './encode';
 import setSS58Format from './setSS58Format';
 
 export {
   checkAddress,
+  checkAddressChecksum,
   decodeAddress,
   encodeAddress,
   setSS58Format